Patch 95_xprint_disable_dbus: disable D-BUS in Xprint
If dbus is enabled in Xprt, it causes Dispatch() to loop without
waiting, using 95% cpu.  Adapting from upstream commit
2a3d1421e0cc18822ae8f478fcc272e16a9e9340, we disable dbus in Xprt
using dummy config functions. We use a patch rather than cherry-pick
because of interference in configure.ac by 94_xprint_XSERVER_LIBS, and
for the same reason share some of the dbus disabling with that patch.

Closes: #472180.