Xi: Fix TouchEnd to TouchUpdate change for one accepted grab
If there is only one listener of a touch, the listener is a grab, and is
accepted before the touch has ended, the current code will not end the
touch record when the touch does end.

This change adds a listener state for when a touch is accepted but has
not yet ended. We now keep the touch record alive in this state, but end
it when the touch ends.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
(cherry picked from commit 58427e08a4a36ce9e213e4b4fe5249d5db2c764d)
(cherry picked from commit 5a3ec826e653377e8b70e7553d1f0ca72210447c)

Conflicts:

	Xi/exevents.c

Signed-off-by: Chase Douglas <chase.douglas@ubuntu.com>