Drop 162_null_crtc_in_rotation.patch as no longer relevant
We've carried this patch for a few years, meantime a number of
improvements have been made to upstream handling of multi-GPU.

We're going to optimistically drop this patch in hopes that whatever
situation led to the crash are long since gone. This bug is kind of a
corner case we don't expect very many people to encounter, so this is a
low risk way of handling it. However, if anyone does see this still
happening, please file a new bug report (and reference this one).
(LP: #139990)