Don't delete outputs without replacing them
This fixes LP: #1504422 and LP: #1602561.

Previously any attempt to resize the root window from fullscreen
resulted in all the RandR outputs getting deleted. Kind of reasonable
since the root window no longer matched the output dimensions.

However it was also unreasonable for three reasons:
  * The behaviour was still wrong for multi-monitor systems.
  * The outputs were deleted and never replaced with anything
    despite a log message to the contrary (LP: #1504422)
  * Some games/toolkits like SDL will fail to start if no outputs
    are found (LP: #1602561)

So now we just report the physical output information and clients
need to be able to cope with that NOT relating to the size of the
root window. But they already had to in -rootless mode so at least
it's consistent now, and two bugs are fixed.